synthetic-data-geneation/pyproject.toml
[project]
name = "synthetic-data-geneation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "ipykernel>=6.29.5",
    "logfire[fastapi]>=3.14.0",
    "openai>=1.72.0",
    "pandas>=2.2.3",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.0",
]
